requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
google-generativeai
pinecone
langchain
langchain-community
pypdf